pyproject.toml
[project]
name = "personal-assistent-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "fabulous>=0.4.0",
    "huggingface-hub>=0.34.4",
    "ipykernel>=6.30.1",
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
]
